refactor(date-picker): component optimization
