+++ packages/components/date-picker/DateRangePickerPanel.tsx
@@ -13,8 +13,8 @@ import {
   DatePickerMonthChangeTrigger,
 } from './type';
 
-import TRangePanel from './panel/RangePanel';
-import useRangeValue from './hooks/useRangeValue';
+import TRangePanel from './components/panel/RangePanel';
+import { useRangeValue } from './hooks/useRangeValue';
 import { formatDate, getDefaultFormat, parseToDayjs } from '@tdesign/common-js/date-picker/format';
 import { subtractMonth, addMonth, extractTimeObj } from '@tdesign/common-js/date-picker/utils';
 import { dateCorrection } from './utils';
