--- packages/components/date-picker/base/style/css.js
@@ -1,1 +1,0 @@
-import '../../style/index.css';
