--- packages/components/date-picker/base/style/index.js
@@ -1,1 +1,0 @@
-import '../../style/index.js';
