+++ packages/components/date-picker/components/base/Cell.tsx
@@ -1,5 +1,5 @@
 import { defineComponent, computed } from 'vue';
-import { usePrefixClass } from '../../hooks/useConfig';
+import { usePrefixClass } from '../../../hooks/useConfig';
 import { extractTimeObj } from '@tdesign/common-js/date-picker/utils';
 import { Dayjs } from 'dayjs';
