+++ packages/components/date-picker/components/base/Table.tsx
@@ -1,11 +1,11 @@
 import { defineComponent, PropType, computed } from 'vue';
 
 import TDatePickerCell from './Cell';
-import { useConfig, usePrefixClass } from '../../hooks/useConfig';
+import { useConfig, usePrefixClass } from '../../../hooks/useConfig';
 import { parseToDayjs } from '@tdesign/common-js/date-picker/format';
 import { isArray } from 'lodash-es';
 
-import type { TdDatePickerProps, DateMultipleValue } from '../type';
+import type { TdDatePickerProps, DateMultipleValue } from '../../type';
 
 export default defineComponent({
   name: 'TDatePickerTable',
