+++ packages/components/date-picker/components/base/style/css.js
@@ -1,0 +1,1 @@
+import '../../../style/index.css';
