+++ packages/components/date-picker/components/base/style/index.js
@@ -1,0 +1,1 @@
+import '../../../style/index.js';
