+++ packages/components/date-picker/components/panel/ExtraContent.tsx
@@ -1,7 +1,7 @@
 import { defineComponent, PropType, computed } from 'vue';
 import TDateFooter from '../base/Footer';
 
-import type { TdDatePickerProps, TdDateRangePickerProps } from '../type';
+import type { TdDatePickerProps, TdDateRangePickerProps } from '../../type';
 
 export default defineComponent({
   name: 'TExtraContent',
