+++ packages/components/date-picker/components/panel/RangePanel.tsx
@@ -1,11 +1,10 @@
 import { defineComponent, PropType, computed } from 'vue';
-import { useConfig, usePrefixClass } from '../../hooks/useConfig';
+import { useConfig, usePrefixClass } from '../../../hooks/useConfig';
 import TPanelContent from './PanelContent';
 import TExtraContent from './ExtraContent';
-import { TdDateRangePickerProps } from '../type';
+import { TdDateRangePickerProps } from '../../type';
 import { getDefaultFormat, parseToDayjs } from '@tdesign/common-js/date-picker/format';
-import useTableData from '../hooks/useTableData';
-import useDisableDate from '../hooks/useDisableDate';
+import { useTableData, useDisableDate } from '../../hooks';
 
 export default defineComponent({
   name: 'TRangePanel',
