+++ packages/components/date-picker/components/panel/style/css.js
@@ -1,0 +1,1 @@
+import '../../../style/index.css';
