+++ packages/components/date-picker/components/panel/style/index.js
@@ -1,0 +1,1 @@
+import '../../../style/index.js';
