+++ packages/components/date-picker/hooks/useTableData.tsx
@@ -2,7 +2,7 @@ import { useConfig } from '../../hooks/useConfig';
 
 import { getWeeks, getYears, getMonths, getQuarters, flagActive } from '@tdesign/common-js/date-picker/utils';
 
-export default function useTableData(props: any) {
+export function useTableData(props: any) {
   // 国际化文本初始化
   const { globalConfig } = useConfig('datePicker');
